stacktraces in report

+++ plugin.sync.bug/src/main/scala/com/anarsoft/race/detection/groupinterleave/GroupInterleaveElement.scala
@@ -4,7 +4,7 @@ import com.anarsoft.race.detection.partialorder.PartialOrderBuilder
 import com.anarsoft.race.detection.stacktrace.StacktraceNode
 
 trait GroupInterleaveElement extends GroupInterleaveElementForResult {
-  def setStacktraceNode(threadIdToStacktraceNodeArray: Map[Long, Array[StacktraceNode]]): Unit;
+  def setStacktraceNode(threadIdToStacktraceNodeArray: Map[Int, Array[StacktraceNode]]): Unit;
 
   def addToPartialOrderBuilder(partialOrderBuilder: PartialOrderBuilder): Unit;
